Move copy_artifacts_dev into dataproc utils
