+++ dags/mozaggregator_prerelease.py
@@ -4,10 +4,9 @@ from datetime import datetime, timedelta
 
 from airflow import DAG
 from airflow.contrib.hooks.gcp_api_base_hook import GoogleCloudBaseHook
-from airflow.operators.bash_operator import BashOperator
 from airflow.operators.moz_databricks import MozDatabricksSubmitRunOperator
 from airflow.operators.subdag_operator import SubDagOperator
-from utils.dataproc import moz_dataproc_pyspark_runner
+from utils.dataproc import moz_dataproc_pyspark_runner, copy_artifacts_dev
 from utils.mozetl import mozetl_envvar
 
 default_args = {
@@ -81,32 +80,6 @@ storage_bucket = (
     if is_dev
     else "moz-fx-data-prod-dataproc-scratch"
 )
-
-
-def copy_artifacts_dev(dag, project_id, artifact_bucket, storage_bucket):
-    return BashOperator(
-        task_id="copy_to_dev_artifacts",
-        bash_command="""
-        gcloud auth activate-service-account --key-file ~/.credentials
-
-        gsutil mb gs://${ARTIFACT_BUCKET}
-        gsutil mb gs://${STORAGE_BUCKET}
-
-        gsutil -m cp -r ~/dataproc_bootstrap gs://${ARTIFACT_BUCKET}
-        gsutil -m cp -r ~/jobs gs://${ARTIFACT_BUCKET}
-
-        echo "listing artifacts..."
-        gsutil ls -r gs://${ARTIFACT_BUCKET}
-        """,
-        env={
-            # https://github.com/GoogleCloudPlatform/gsutil/issues/236
-            "CLOUDSDK_PYTHON": "python",
-            "PROJECT_ID": project_id,
-            "ARTIFACT_BUCKET": artifact_bucket,
-            "STORAGE_BUCKET": storage_bucket,
-        },
-        dag=dag,
-    )
 
 
 prerelease_telemetry_aggregate_view_dataproc = SubDagOperator(
